requirements.txt
psutil>=5.9.0
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
tabulate>=0.9.0
colorama>=0.4.6
rich>=13.0.0
schedule>=1.2.0
pyyaml>=6.0
sqlite3
argparse
datetime
json
csv
os
platform
socket
urllib.parse
winreg
subprocess
threading
time
logging 